moving logging configuration to HARK.__init__
